refactor: make the fixture overridable
