+++ tests/conftest.py
@@ -11,17 +11,12 @@ def pytest_configure():
 
 
 @pytest.fixture(scope="session")
-def test_folder(gee_hash, account_root):
-    """Create a test folder for the test session."""
+def gee_folder_structure():
+    """Override the default test folder structure."""
     point = ee.Geometry.Point([0, 0])
-    structure = {
+    return {
         "folder": {
             "image": ee.Image(1).clipToBoundsAndScale(point.buffer(100), scale=30),
             "fc": ee.FeatureCollection(point),
         }
     }
-    folder = pytest_gee.init_tree(structure, gee_hash, account_root)
-
-    yield folder
-
-    pytest_gee.delete_assets(folder, False)
