+++ tests/test_pytest_gee.py
@@ -15,19 +15,27 @@ def test_gee_init():
     assert ee.Number(1).getInfo() == 1
 
 
-def test_init_tree(account_root, test_folder):
+def test_structure(gee_folder_structure):
+    """Test the structure fixture."""
+    assert isinstance(gee_folder_structure, dict)
+    assert "folder" in gee_folder_structure
+    assert "image" in gee_folder_structure["folder"]
+    assert "fc" in gee_folder_structure["folder"]
+
+
+def test_init_tree(gee_folder_root, gee_test_folder):
     """Test the init_tree function."""
     # search all the assets contained in the test_folder
-    asset_list = pytest_gee.utils.get_assets(account_root)
+    asset_list = pytest_gee.utils.get_assets(gee_folder_root)
     asset_list = [i["name"] for i in asset_list]
 
     # identify specific files and folders
-    folder = test_folder / "folder"
+    folder = gee_test_folder / "folder"
     image = folder / "image"
     feature_collection = folder / "fc"
 
     # check that they exist
-    assert str(test_folder) in asset_list
+    assert str(gee_test_folder) in asset_list
     assert str(folder) in asset_list
     assert str(image) in asset_list
     assert str(feature_collection) in asset_list
